extract meta api
