+++ common/meta-apis/meta-api-vo/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "common-kv-api"
+name = "common-meta-api-vo"
 version = "0.1.0"
 authors = ["Databend Authors <opensource@datafuselabs.com>"]
 license = "Apache-2.0"
@@ -10,11 +10,7 @@ edition = "2021"
 
 [dependencies]
 
-common-metatypes = {path= "../../metatypes" }
-common-exception = {path = "../../exception" }
-common-store-api-util= { path = "../util" }
-common-kv-api-vo= { path = "../kv-api-vo" }
+common-metatypes = {path= "../../metatypes"}
+common-datavalues= {path= "../../datavalues"}
 
-
-futures = "0.3"
-async-trait = "0.1"
+serde = { version = "1.0", features = ["derive"] }
