+++ metasrv/Cargo.toml
@@ -72,6 +72,7 @@ axum-server = { version = "0.2", features = ["tls-rustls"] }
 [dev-dependencies]
 
 common-kv-api = {path = "../common/kv-apis/kv-api" }
+common-meta-api = {path = "../common/meta-apis/meta-api" }
 
 pretty_assertions = "0.7"
 test-env-log = "0.2.7"
