+++ query/src/common/store_api_provider.rs
@@ -17,7 +17,7 @@ use std::sync::Arc;
 
 use common_exception::Result;
 use common_kv_api::KVApi;
-use common_store_api::MetaApi;
+use common_meta_api::MetaApi;
 use common_store_api::StorageApi;
 use common_store_api_sdk::StoreClient;
 use common_store_api_sdk::StoreClientConf;
