+++ store/src/api/rpc/flight_service_test.rs
@@ -19,6 +19,8 @@ use std::time::UNIX_EPOCH;
 use common_datablocks::DataBlock;
 use common_datavalues::prelude::*;
 use common_kv_api::KVApi;
+use common_meta_api::MetaApi;
+use common_meta_api_vo::*;
 use common_metatypes::KVMeta;
 use common_metatypes::KVValue;
 use common_metatypes::MatchSeq;
@@ -28,9 +30,6 @@ use common_planners::DropDatabasePlan;
 use common_planners::DropTablePlan;
 use common_planners::ScanPlan;
 use common_runtime::tokio;
-use common_store_api_sdk::meta_api_impl::DropTableActionResult;
-use common_store_api_sdk::meta_api_impl::GetTableActionResult;
-use common_store_api_sdk::MetaApi;
 use common_store_api_sdk::StorageApi;
 use common_store_api_sdk::StoreClient;
 use common_tracing::tracing;
