+++ store/src/executor/action_handler_test.rs
@@ -20,6 +20,7 @@ use common_datavalues::DataField;
 use common_datavalues::DataSchema;
 use common_datavalues::DataType;
 use common_exception::ErrorCode;
+use common_meta_api_vo::*;
 use common_planners::CreateDatabasePlan;
 use common_planners::CreateTablePlan;
 use common_planners::DropDatabasePlan;
@@ -28,17 +29,11 @@ use common_runtime::tokio;
 use common_runtime::tokio::sync::mpsc::Receiver;
 use common_runtime::tokio::sync::mpsc::Sender;
 use common_store_api_sdk::meta_api_impl::CreateDatabaseAction;
-use common_store_api_sdk::meta_api_impl::CreateDatabaseActionResult;
 use common_store_api_sdk::meta_api_impl::CreateTableAction;
-use common_store_api_sdk::meta_api_impl::CreateTableActionResult;
 use common_store_api_sdk::meta_api_impl::DropDatabaseAction;
-use common_store_api_sdk::meta_api_impl::DropDatabaseActionResult;
 use common_store_api_sdk::meta_api_impl::DropTableAction;
-use common_store_api_sdk::meta_api_impl::DropTableActionResult;
 use common_store_api_sdk::meta_api_impl::GetDatabaseAction;
-use common_store_api_sdk::meta_api_impl::GetDatabaseActionResult;
 use common_store_api_sdk::meta_api_impl::GetTableAction;
-use common_store_api_sdk::meta_api_impl::GetTableActionResult;
 use common_store_api_sdk::storage_api_impl::AppendResult;
 use common_store_api_sdk::storage_api_impl::TruncateTableAction;
 use common_store_api_sdk::storage_api_impl::TruncateTableResult;
