+++ store/src/executor/meta_handlers.rs
@@ -22,6 +22,7 @@ use common_arrow::arrow::io::ipc::write::common::IpcWriteOptions;
 use common_arrow::arrow_flight::utils::flight_data_from_arrow_schema;
 use common_arrow::arrow_flight::FlightData;
 use common_exception::ErrorCode;
+use common_meta_api_vo::*;
 use common_metatypes::Cmd::CreateDatabase;
 use common_metatypes::Cmd::CreateTable;
 use common_metatypes::Cmd::DropDatabase;
@@ -31,20 +32,12 @@ use common_metatypes::LogEntry;
 use common_metatypes::Table;
 use common_raft_store::state_machine::AppliedState;
 use common_store_api_sdk::meta_api_impl::CreateDatabaseAction;
-use common_store_api_sdk::meta_api_impl::CreateDatabaseActionResult;
 use common_store_api_sdk::meta_api_impl::CreateTableAction;
-use common_store_api_sdk::meta_api_impl::CreateTableActionResult;
-use common_store_api_sdk::meta_api_impl::DatabaseMetaReply;
-use common_store_api_sdk::meta_api_impl::DatabaseMetaSnapshot;
 use common_store_api_sdk::meta_api_impl::DropDatabaseAction;
-use common_store_api_sdk::meta_api_impl::DropDatabaseActionResult;
 use common_store_api_sdk::meta_api_impl::DropTableAction;
-use common_store_api_sdk::meta_api_impl::DropTableActionResult;
 use common_store_api_sdk::meta_api_impl::GetDatabaseAction;
-use common_store_api_sdk::meta_api_impl::GetDatabaseActionResult;
 use common_store_api_sdk::meta_api_impl::GetDatabaseMetaAction;
 use common_store_api_sdk::meta_api_impl::GetTableAction;
-use common_store_api_sdk::meta_api_impl::GetTableActionResult;
 use common_store_api_sdk::meta_api_impl::GetTableExtReq;
 use log::info;
